clean up duplicate types
