+++ ts/packages/commonUtils/src/command.ts
@@ -6,71 +6,6 @@ export type Command = {
     args?: string[];
 };
 
-// TODO: remove duplicate types due to package circular dependencies (shell/electronTypes.ts is other source)
-export type SearchMenuItem = {
-    matchText: string;
-    selectedText: string;
-    emojiChar?: string;
-    groupName?: string;
-};
-
-export type ActionInfo = {
-    name: string;
-    comments: string;
-    item: SearchMenuItem;
-    template?: ActionTemplate | undefined;
-};
-
-export type TemplateParamPrimitive = {
-    type: "string" | "number" | "boolean";
-};
-
-export type TemplateParamStringUnion = {
-    type: "string-union";
-    typeEnum: string[];
-};
-
-export type TemplateParamScalar =
-    | TemplateParamPrimitive
-    | TemplateParamStringUnion;
-
-export type TemplateParamArray = {
-    type: "array";
-    elementType: TemplateParamField;
-};
-
-export type TemplateParamObject = {
-    type: "object";
-    fields: {
-        [key: string]: TemplateParamFieldOpt;
-    };
-};
-
-export type TemplateParamFieldOpt = {
-    optional?: boolean;
-    field: TemplateParamField;
-};
-
-export type TemplateParamField =
-    | TemplateParamScalar
-    | TemplateParamObject
-    | TemplateParamArray;
-
-export type ActionTemplate = {
-    agent: string;
-    name: string;
-    parameterStructure: TemplateParamObject;
-};
-
-export type ActionTemplateSequence = {
-    templates: ActionTemplate[];
-    actions: unknown;
-    preface?: string;
-    editPreface?: string;
-};
-
-// end duplicate type section
-
 export function parseCommandLine(line: string): Command | undefined {
     if (line.length == 0) {
         return undefined;
